requirements.txt
pxr
paho-mqtt
